Text from PDF is extracted and written to a txt file
