+++ python/data_preprocessing/output/test.txt
@@ -1,0 +1,1 @@
+1. Ubung: EPK/BPMN-Modellierung  Aufgabe 1 Modellieren Sie folgenden Prozessablauf als EPK (und behalten Sie ihn fur die nachste Party ;-)).  Wenn ein Gast einen Pina Colada bestellt, achtest du als erstes darauf, dass das Glas in der Gefriertruhe steht, sodass es die richtige Temperatur bekommt. Parallel dazu gibst du 4 Eiswirfel in einen Shaker. Zudem fugst du 6cl Rum, 2cl Sahne, 4cl Kokoslikér und 12cl Ananassaft dazu. Dann den Shaker verschlieRen und anschlieRend alles fur 10 Sekunden schttteln. Wenn du mit dem Shaken fertig bist und das Glas kalt ist, gibst du die Mischung aus dem Shaker in das Glas. Im nachsten Schritt prifst du, ob Ananas-Sticke vorhanden sind. Wenn ja, nimmst du zwei Stucke, um den Drink zu garnieren. Wenn nein, erhalt der Drink keine Dekoration. AnschlieRend servierst du den Drink dem glticklichen Kunden. Zimmer staubsaugen  Zimmer gestaubsaugt  Pizza abholen  Pizza abholen  Freunde eingeladen  Tinder-Match eingeladen  Freundin eingeladen  Wohnung  Bierpfand dekorieren  wegpbringen  Freunde abholen Netflix ver  rlangern Tinder l6schen  Freunde abgeholt  Wohnung dekoriert  Netflix verlangert  Bierpfand weggebracht  Person(en) begriiRen  Person(en) der Kasse  Kunde wartet an  Kundenanruf kommt an  Bestellung angelegt  Bestellung ist eingepflegt  Gesamtpreis berechnen  Rechnung erstellt  Kunde wahit Bezahloption  abwickeln  Kauf ist abgeschlossen  Bargeldzahlung  Zahlung mit Kreditkarte gewahit  Bezahlsystem  Zahlungsfahig- keit ermittelt  Unterschrift prifen  Ablaufdatum priifen  Daten sind vollstandig eine Anfrage bei einem Hotel mit der Anzahl der positiven Ruckmeldungen. Das Hotel pruft nun die Anfrage. Sind keine Zimmer mehr verfiigbar, erhalt Michael eine Absage. Dieser sucht darauf ein neues Hotel und sendet eine weitere Anfrage. Sind im initial angefragtem Hotel Zimmer frei, erhalt Michael ein Angebot fiir den Urlaub. Michael gibt die Eckpunkte zum Ziel und den Kosten an seine Familie weiter. Die Familie Uberweist dann das Geld und Michael wartet bis er das Geld von allen erhalten hat. Dann Uberweist er den Betrag an die Unterkunft. Das Hotel sendet daraufhin eine Buchungsbestatigung an Michael. Der Urlaub kann dann starten.
